Allow WayfException as cause of WayfRequestHandled
Allow passing a WayfException as the cause of a WayfRequestHandled
exception.

This allows to signal an error from inside a plugin
(SamlIdpCookiePlugin).

When forwardRequest encounters an error (e.g., protocol requested -
SAML1 - not supported by IdP), the DS would otherwise return a blank
screen when this is done via the redirect cookie (or alwaysFollow).

This change allow to handle this gracefully (i.e., redirect to the error
page with the appropriate error message displayed).

As part of the changes, we had to add constructors to WayfRequestHandled
(to pass a cause of the exception).