Config: Keep IdP siteSets in the config file order
Solves \#8

Modify the config file parsing in WayfService to preserve the order in
which federations (MetadataProvider stored as IdPSiteSet) are listed in
the configuration file.

This means switching from a Hashtable to a List - also in the arguments
passed to the DiscoveryServiceHandler constructor.  Inside the
constructor, construct on-the-fly a Hashtable for looking up IdPSiteSets
by their identifier but add them to the internal config for the handler
from the list, preserving their order.

Also in DSH.setupReturnAddress, when looking up the SP by spName, break
out of the loop after first match to successfully implement the logic
requested in \#8.