DSH: Get IP from X-Forwarded-For header for session logging
Introduce a feature that allows the session logging part of the DSH to get the
client IP address from the X-Forwarded-For header - instead of relying just on
req.getRemoteHost().

This feature is needed to use the DS behind load-balancers that would otherwise
hide the client IP address.

This feature is controlled by useForwardedFor boolean attribute on the Default
element - and is off by default.

Changes done:
* add the useForwardedFor attribute to default wayfconfig.xml
* add the attribute and parsing code to HandlerConfig
* change the session logging code in DSH.forwardRequest() to:
(1) check whether this feature is on
(2) and a header is present
(3) and the header value syntactically matches an IP address.
(4) and in that case, use the header value
(5) and in all other cases, use req.getRemoteHost()