DSH: Fix interactino with SamlCookiePlugin
The changes introduced in 3b88ce74cb change the method signature for
forwardRequest - changing it from public static to private.

This overlooks that
edu.internet2.middleware.shibboleth.wayf.plugins.provider.SamlCookiePlugin
is also invoking this method.

The method therefore needs to be public: so reverting from private back to public.

To give the SamlCookiePlugin a way to invoke this method in non-static
context, add a parameter pointing back to the DSH to the signature of
all methods in the Plugin interface.  (As a plugin instance can be used
by multiple DSH instances, there is no way for a plugin to hold a
permanent reference to the DSH - so it must be passed in each
invocation.  Sorry to be changing an interface....)

Fixes #11