Safari scrolling fix
We need to implement a work-around for broken SELECT list scrolling in
Safari.

This is a known bug in Safari 8 - documented e.g. at
http://stackoverflow.com/questions/27109190/safari-8-multiple-select-scrolling-issue

For SELECT lists that have a SIZE attribute specified, Safari does not
scroll the list when the user gets past the last visible element (e.g.,
via using the down arrow or typing the initial letter of an entry in the
list).

The solution described on the above link only handles the up and down
keys and no other methods of selection (e.g., typing initial letters of
name), so instead of hooking into the keyboard events, we are listening
for the change event - adding the function as an onchange hook.

Note also that the scrollByLines method is officially defined only for
Window, not abitrary (SELECT) element.
And consequently this method is visible on Safari but not on other
browsers (and results into JS errors).  Therefore, the code checks for
presence of the scrollByLines method and only engages if this method is
present.  Which means it may only act on Safari - but, other browsers do
not have the scrolling bug, so, no action is required there.

Also note that Safari jumps to the top of the list when making a
selection ABOVE the currently visible part of the list.

This may cause some erratic scrolling behaviour when the user is going
through the list UP one-by-one - when reaching the top of the screen,
Safari automatically jumps to the top, and this function may instead end
up going down - if the top position Safari scrolled to is too high for
the entry selected.

And finally note, we need to initialize the scrolling fix by populating
the scrollLineHeight variable - by attempting to scroll one line down
and up again and measuring the pixel distance - needed as some scrolling
information is measured in pixels and some in lines.

And as the scrolling can only be done with a list long enough to scroll,
we have to reinitialize whenever the list changes (when a different
federation is selected) - hence the multiple calls to
refreshSrollParams.

And all of the above applies only when the DS is run in the ListOfLists
mode - the single list displayed otherwise does not have a size element
and thefore this scrolling issue is not triggered.  And in that case, we
do not render the code - it is only active when siteLists is present in
the request.